+++ openquake/hazardlib/gsim/base.py
@@ -28,6 +28,7 @@ import functools
 import numpy
 from scipy.special import ndtr
 
+from openquake.hazardlib.stats import norm_cdf
 from openquake.hazardlib.probability_map import ProbabilityCurve
 from openquake.baselib.general import DeprecationWarning
 from openquake.hazardlib import imt as imt_module
@@ -218,31 +219,6 @@ def _get_poes(mean_std, loglevels, truncation_level, squeeze=False):
                 out[:, lvl] = (iml - mean[:, m]) / stddev[:, m]
             lvl += 1
     return _truncnorm_sf(truncation_level, out)
-
-
-from scipy.stats import norm
-def norm_cdf(x, a, s):
-    """
-    Gaussian cumulative distribution function; if s=0, returns an
-    Heaviside function instead. NB: for x=a, 0.5 is returned for all s.
-
-    >>> norm_cdf(1.2, 1, .1)
-    0.9772498680518208
-    >>> norm_cdf(1.2, 1, 0)
-    1.0
-    >>> norm_cdf(.8, 1, .1)
-    0.022750131948179216
-    >>> norm_cdf(.8, 1, 0)
-    0.0
-    >>> norm_cdf(1, 1, .1)
-    0.5
-    >>> norm_cdf(1, 1, 0)
-    0.5
-    """
-    if s == 0:
-        return numpy.heaviside(x - a, .5)
-    else:
-        return norm.cdf(x, loc=a, scale=s)
 
 
 def _get_poes_site(mean_std, loglevels, truncation_level, ampfun,
