+++ openquake/hazardlib/site_amplification.py
@@ -20,7 +20,8 @@ import re
 import copy
 import numpy
 import pandas as pd
-from scipy.stats import norm
+
+from openquake.hazardlib.stats import norm_cdf
 from openquake.hazardlib.site import ampcode_dt
 from openquake.hazardlib.imt import from_string
 from openquake.hazardlib.probability_map import ProbabilityCurve
@@ -47,7 +48,7 @@ class AmplFunction():
         self.df = df
 
     @classmethod
-    def from_compact(cls, df, soil=None):
+    def from_dframe(cls, df, soil=None):
         """
         :param df:
             A :class:`pandas.DataFrame` instance
@@ -127,33 +128,9 @@ class AmplFunction():
         :returns:
             The maximum sigma value in the amplification function
         """
-        #return max(self.df['std'])
+        # TODO - this must be improved
+        # return max(self.df['std'])
         return 0.3
-
-
-# this is necessary since norm.cdf for scale=0 returns NaNs
-def norm_cdf(x, a, s):
-    """
-    Gaussian cumulative distribution function; if s=0, returns an
-    Heaviside function instead. NB: for x=a, 0.5 is returned for all s.
-
-    >>> norm_cdf(1.2, 1, .1)
-    0.9772498680518208
-    >>> norm_cdf(1.2, 1, 0)
-    1.0
-    >>> norm_cdf(.8, 1, .1)
-    0.022750131948179216
-    >>> norm_cdf(.8, 1, 0)
-    0.0
-    >>> norm_cdf(1, 1, .1)
-    0.5
-    >>> norm_cdf(1, 1, 0)
-    0.5
-    """
-    if s == 0:
-        return numpy.heaviside(x - a, .5)
-    else:
-        return norm.cdf(x, loc=a, scale=s)
 
 
 def check_unique(df, kfields, fname):
